requirements.txt
asgiref==3.3.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.1.8
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.28.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
openapi-codec==1.3.2
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
pytz==2021.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
simplejson==3.20.1
sqlparse==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
